checkout view description
So I made use of the gridpane layout to make 4 grids around the console and then I made separate gridpanes for the  checkout and the payment info part. I added the checkbox button to the payment part and the login button for the checkout part after this I added 2 vboxes for the other side of the gridpane. I added the required button in that side of the gridpane and made use of an HBox to contain the couponText and its textField.